test: add cross-field syntax tests
